refactor: split optimizers.py into optimizers/ subpackage
optimizers/
├── core.py    # _OptimizationState, _common_setup, _batched_project
├── gd.py      # RiemannianGD dataclass + _armijo_step
├── adam.py    # RiemannianAdam dataclass + _init_adam_state + _adam_step
└── loop.py    # optimize() dispatcher

Public API unchanged: from pdft.optimizers import RiemannianGD,
RiemannianAdam, AbstractRiemannianOptimizer, optimize  still works
via __init__.py re-exports. The Adam-specific JIT fast path used by
batched training (training/adam_step.py, next commit) is intentionally
a separate implementation — see comment in optimizers/adam.py.

Parity tests (test_parity_adam.py, test_parity_long_run.py) pass.
